Phase03/backend/pyproject.toml
# [Task T002] Backend Python project configuration with UV

[project]
name = "todo-backend"
version = "0.1.0"
description = "Todo Full-Stack Web Application - FastAPI Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "sqlmodel>=0.0.14",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyjwt>=2.8",
    "uvicorn[standard]>=0.27",
    "asyncpg>=0.29",
    "python-dotenv>=1.0",
    "openai>=1.0.0",
    "openai-agents>=0.1.0",
    "mcp>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.scripts]
uvicorn = "uvicorn.main:main"

[dependency-groups]
dev = []
